backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.0.0
python-dotenv==1.0.0
